Defensive programming.

+++ grow-array.h
@@ -2,10 +2,11 @@
 #define _GROW_ARRAY_H_
 
 #include <stdint.h>
+#include <stdlib.h>
 
-void *grow_array_using_realloc(void *a, uint32_t *size, uint32_t target_size);
+void *grow_array_using_realloc(void *a, uint32_t *size, size_t target_size);
 
-static inline void *grow_array(void *a, uint32_t *size, uint32_t target_size)
+static inline void *grow_array(void *a, uint32_t *size, size_t target_size)
 {
     uint32_t n = *size;
     if (target_size <= n)
